backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loom-backend"
version = "0.1.0"
description = "Loom — local-first AI memory system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "uvicorn[standard]>=0.30",
    "watchdog>=4.0",
    "openai>=1.0",
    "anthropic>=0.39",
    "httpx>=0.27",
    "lancedb>=0.15",
    "pyarrow>=17.0",
    "tiktoken>=0.7",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["api", "agents", "index", "core"]

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true
explicit_package_bases = true
exclude = ["tests/"]
